+++ .eslintrc.cjs
@@ -1,9 +1,0 @@
-/**
- * @type {import('eslint').Linter.Config}
- */
-module.exports = {
-    env: {
-        node: true,
-    },
-    extends: '.',
-};